app_root/requirements.txt
openai
chromadb
tiktoken
python-dotenv
rich
typer
pydantic
fastapi
uvicorn